feat:(modularizacion total de la pagina de edificios)
se tomo el codigo completo original de la pagina de edificios y se termino su modilarizacion aunque actualmente el unico problema de esto seria la coneccion con la api por que actualmente creo que no existe endpoits para esta pagina
